feat: merge exporter and data_registry apps
