+++ data_registry/urls.py
@@ -6,6 +6,7 @@ urlpatterns = [
     path("", views.index, name="index"),
     path("search/", views.search, name="search"),
     path("publication/<int:id>", views.detail, name="detail"),
+    path("publication/<int:id>/download", views.download_export, name="download"),
     # https://code.djangoproject.com/ticket/26556
     path("i18n/setlang/", i18n.set_language, name="set-language"),
     # Uncomment after re-integrating Spoonbill.
